requirements.txt
streamlit==1.42.2
openai==1.109.1
anthropic==0.46.0
azure-storage-blob==12.25.1
